+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0-next.104](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.103...@esri/calcite-components-react@3.0.0-next.104) (2025-01-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.103](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.102...@esri/calcite-components-react@3.0.0-next.103) (2025-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.102](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.101...@esri/calcite-components-react@3.0.0-next.102) (2025-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.101](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.100...@esri/calcite-components-react@3.0.0-next.101) (2025-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.0.0-next.100](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.99...@esri/calcite-components-react@3.0.0-next.100) (2025-01-14)
 
 **Note:** Version bump only for package @esri/calcite-components-react